+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   (used for the log `message`) and `event.dataset` (a field provided by the
   `elasticapm` integration) ([#46](https://github.com/elastic/ecs-logging-python/pull/46))
 * Add default/fallback handling for json.dumps ([#47](https://github.com/elastic/ecs-logging-python/pull/47))
+* Fixed an issue in `StdLibFormatter` when `exc_info=False` ([#42](https://github.com/elastic/ecs-logging-python/pull/42))
 
 ## 1.0.0 (2021-02-08)